build: switch from alpine to slim image (#299)
investigating steady increase in memory consumption, which can be
observed in rancher and which leads to regular container restarts caused
by OOM.

we are definitely leaking sockets - `lsof -p 1 | wc -l` showed up to a
million!

running a frontend container locally, it looks as if this issue does not
occur when switching from `alpine` to a regular `glibc` based `node`
image. testing that theory with this PR.

since image optimisation is still disabled we should not run into memory
issues caused by `sharp`, which is the actual reason we are using
`alpine` images.